+++ runtime/src/f2j/unbox/ClosureIntBox.java
@@ -2,7 +2,7 @@ package f2j.unbox;
 
 public abstract class ClosureIntBox implements Closure
 {
-  public int x;
+  public long x;
   public Object out;
   public abstract void apply ()
   ;
